Moved CashNetBRepository contents to TransactionRepository
